Remove usage of Addr in contract messages.
The Addr type is not validated during deserialization yet it still
implements both serialization and deserialization. This means that an
invalid Addr can be constructed in the event that it is passed to a
contract via a message.

This removes all usage of Addr in message types. Before this change
two bugs are present:

1. For the DAO and Multisig contracts tokens can be added to the
treasury which are invalid.
2. For the staking contract this allows setting the admin of the
contract to an invalid address.